refactoring unnecessary files; defining all types in one file for simplicity and eloquency
