--- src/Data/Main.hs
@@ -1,12 +1,0 @@
-module Data.Main where
-
-import Data.FM.Feature
-import Data.FM.FeatureTree
-import Data.FM.FeatureModel
-import Data.FM.Expression
-import Data.FM.SAT
-import Data.FM.Utils
-
-
-main :: IO ()
-main = putStrLn "someFunc"
