+++ test/Spec.hs
@@ -1,5 +1,3 @@
-import Test.FM.FeatureTest
-import Test.FM.TreeTest
 import Test.FM.ExpressionTest
 import Test.FM.FeatureModelTest
 import Test.FM.ProductConfigurationTest
@@ -9,9 +7,8 @@ import Control.Monad
 import Test.HUnit.Base
 
 main :: IO Test.HUnit.Base.Counts
--- main = runFeatureTests
 main = do
-    a    <- runFeatureTests
+    a    <- runExpressionTests
     b    <- runFeatureModelTests
     c    <- runProductConfigurationTests
     -- d    <- runXMLParserTests
